Fix two silent numeric differences the usage-limit suite exposed
SUM over a BIGINT column returns numeric, not int8 — Postgres widens it so the
sum cannot overflow — and `pg` returns numeric as a string for the same reason
it does int8. The analytics code adds those sums, compares them and sends them
to the charts, so a token total rendered as "600" and two of them added up to
"600200". No column in this schema is declared NUMERIC, so the parser only
ever sees an aggregate of values that were already safe as numbers.

The second is worse, because nothing errors. SQLite's strftime('%s', ...)
yields an integer, so `strftime('%s', t) / 900` truncates. EXTRACT returns
numeric, where the division is exact and the surrounding CAST rounds instead —
a row recorded at 10:11 landed in the 10:15 bucket where SQLite put it in
10:00, and two rows that used to aggregate into one stopped doing so. The
timeline would simply have shown different numbers.

Porting the suite also turned up why it could not have caught either: it
declared its own SQLite schema with no foreign keys, so it inserted usage rows
for sessions that did not exist. It now runs against the real schema, where
Postgres refuses them.

+++ packages/backend/src/db/dialect.ts
@@ -156,7 +156,13 @@ function translateStrftime(args: string): string | null {
   let value = `(${column})::timestamp`;
   if (modifier) value = `(${value} + (${modifier})::interval)`;
 
-  if (format === '%s') return `EXTRACT(EPOCH FROM ${value})`;
+  // `::bigint`, and it matters. SQLite's strftime('%s', ...) yields an integer,
+  // so `strftime('%s', t) / 900` is integer division and truncates. EXTRACT
+  // returns numeric, where the same division is exact and the surrounding CAST
+  // rounds instead — which puts a row recorded at 10:11 into the 10:15 bucket
+  // while SQLite put it in 10:00. Two rows that used to aggregate into one
+  // silently stop doing so.
+  if (format === '%s') return `EXTRACT(EPOCH FROM ${value})::bigint`;
 
   const pattern = translateStrftimeFormat(format!);
   return pattern === null ? null : `to_char(${value}, '${pattern}')`;
